extend elm-webapp executable with "crud"
"crud" is slightly different from other cli options
e.g. "elm-webapp crud Post blog"

will patch a crud feature for "Post" into "./blog" directory if "./blog" already exists

but will generate a "./blog" application first, if "./blog" does not exist yet
