+++ package.json
@@ -1,6 +1,5 @@
 {
   "bin": {
-    "elm-webapp-scaffold": "./bin/elm-webapp-scaffold.js",
     "elm-webapp": "./bin/elm-webapp.js"
   },
   "name": "elm-webapp",
